+++ root/LuckyItemDAO.java
@@ -1,3 +1,5 @@
+package root;
+
 import java.io.BufferedReader;
 import java.io.File;
 import java.io.FileInputStream;
@@ -7,11 +9,11 @@ import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
 
-public class FortuneResultDAO{
+public class LuckyItemDAO{
     // 星座一覧を格納した配列
     private final List<String> SIGNS = Arrays.asList("やぎ座","みずがめ座","うお座","おひつじ座","おうし座","ふたご座","かに座","しし座","おとめ座","てんびん座","さそり座","いて座");
     // 占い結果を格納したテキストファイルを指定
-    private final File FILE = new File("FortuneResult.txt");
+    private final File FILE = new File("text/LuckyItemList.txt");
     
         
     public String execute(String sign){
@@ -20,14 +22,14 @@ public class FortuneResultDAO{
                 return null;
             }
             // 占い結果テキストの中身を格納する配列
-            List<String> fortuneResult = new ArrayList<String>();
+            List<String> luckyItemList = new ArrayList<String>();
             // テキストの読み込みをして配列に格納
             try{
                 BufferedReader br = new BufferedReader(new InputStreamReader(new FileInputStream(FILE),"UTF-8"));
                 
                 String line;
                 while((line = br.readLine()) != null){
-                    fortuneResult.add(line);
+                    luckyItemList.add(line);
                 }
                 // 実行した日（今日）を生成
                 LocalDate today = LocalDate.now();
@@ -38,9 +40,9 @@ public class FortuneResultDAO{
                 // 受け取った星座のindex番号を配列から取得
                 int indexOfSIGNS = SIGNS.indexOf(sign);
                 // index番号と今日の日付を使って占い結果のindex番号を生成
-                int targetNum = (y * m / d * (indexOfSIGNS + 1)) % (fortuneResult.size());
+                int targetNum = (y * m / d * (indexOfSIGNS + 1)) % (luckyItemList.size());
                 // 占い結果の配列からindex番号のものを取得
-                String result = fortuneResult.get(targetNum);
+                String result = luckyItemList.get(targetNum);
                 return result;
                 
             } catch (Exception e){
