+++ test/TestGetSignLogic.java
@@ -1,4 +1,6 @@
+package test;
 
+import root.GetSignLogic;
 
 public class TestGetSignLogic {
 
@@ -8,8 +10,8 @@ public class TestGetSignLogic {
         int[] birthday = new int[2];
         String sign;
 
-        // 正常系
-        // 誕生日が適正な数字だった時、星座がかえってくる
+        System.out.println("正常系"); 
+        System.out.println("誕生日が適正な数字だった時、星座がかえってくる"); 
         birthday[0] = 1;
         birthday[1] = 1;
         sign = gsl.getSign(birthday);
@@ -18,8 +20,9 @@ public class TestGetSignLogic {
         } else {
             System.out.println("NG");
         }
-        // 異常系
-        // 誕生日が不正な数字だった時、星座はnullになる
+
+        System.out.println("異常系"); 
+        System.out.println("誕生日が不正な数字だった時、星座はnullになる"); 
         birthday[0] = 0;
         birthday[1] = 0;
         sign = gsl.getSign(birthday);
@@ -28,7 +31,8 @@ public class TestGetSignLogic {
         } else {
             System.out.println("NG");
         }
-        // 誕生日がnullだった時、星座はnullになる
+
+        System.out.println("誕生日がnullだった時、星座はnullになる"); 
         birthday[0] = 0;
         birthday[1] = 0;
         sign = gsl.getSign(birthday);
