+++ text/LuckyItemList.txt
